refactor(game): relocate effectful render to lib/effects (T32)
The layer-boundary lint rule introduced in T31 surfaced one
violation: effects/loop.ts importing ../render/render.js. Since
render.ts directly calls context.fillRect (a Canvas side-effect),
A-3 places it in effects/ until the pure projection layer is
reintroduced as RenderCommand producers in Phase 7. Moved the
module and its DOM-touching test alongside the rest of the shell;
loop.ts now resolves render via a sibling import. Restored the
render/.gitkeep placeholder for the future pure layer.
