+++ packages/game/lib/effects/render.ts
