+++ specs/codebase-compliance/tasks.md
@@ -50,7 +50,7 @@ Each task ends in green `pnpm run ok` for the affected package(s)
 - [x] T29 — Move `packages/game/lib/render.ts` → `packages/game/lib/render/render.ts`. Update importers. `render.test.ts` moved alongside its source; the only outside importer was `lib/loop.ts`, updated to `./render/render.js`. `render/.gitkeep` removed.
 - [x] T30 — Move `packages/game/lib/loop.ts`, `curtain-up.ts`, `bruff-game.ts`, `observable/` → `packages/game/lib/effects/` (preserving the `observable/` subdirectory). Update importers. Two test files moved alongside their sources (`curtain-up.test.ts` and the three `observable/*.test.ts`). `loop.ts` imports of state/render/core became `../...`; `observable/touch.ts` const import became `../../core/constants.js` (one level deeper). Package entry path updated in `package.json` (`main`, `exports.import`) and `vite.config.lib.ts` (`build.lib.entry`); `vitest.config.ts` coverage `exclude` paths updated to the new `effects/...` locations. After this commit `lib/` contains only the five layer directories — no top-level files.
 - [x] T31 — Add per-layer `no-restricted-imports` rule blocks to `@bruff/eslint-config` covering all five layers. Each block lists only the forbidden upward neighbours. Five flat-config blocks appended to `bruff-lint-typescript.js`, each scoped via `files: ["**/lib/<layer>/**/*.ts"]` (so the rule only fires where a layered package adopts the convention). Patterns reference the directory name (e.g. `**/render/**`) so they match the relative-path import strings actually written in source. Sanity-checking the rule against `@bruff/game` surfaces exactly one expected violation — `effects/loop.ts` → `../render/render.js` — which T32 will resolve.
-- [ ] T32 — Run `pnpm run lint` from repo root; resolve any violation by relocating the offending file, not by widening the rule.
+- [x] T32 — Run `pnpm run lint` from repo root; resolve any violation by relocating the offending file, not by widening the rule. The lone violation was `effects/loop.ts` → `../render/render.js`. Diagnosis: the offending file is `render.ts` itself — it directly calls `context.fillRect`, a Canvas side-effect, so under A-3 it belongs in `effects/` until the pure projection layer is reintroduced as `RenderCommand` producers in Phase 7. Moved `render.ts` and `render.test.ts` (which uses `document.createElement` and so is also DOM-bound, per GT-2) to `lib/effects/`; `loop.ts` now imports `./render.js` from a sibling. Restored `lib/render/.gitkeep` so the directory survives until T46/Phase 7 repopulates it. **Spec divergence**: T46 (Phase 9) and T58 reference `lib/render/render.ts`; both will need to track the new path or follow the upcoming `RenderCommand` refactor — to be reconciled per S-25 in T58.
 
 ## Phase 7 — Action taxonomy + reducers
